api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "brain api"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.124.0",
    "psycopg2-binary>=2.9.11",
]
